Adjust declarations to allow "better " encapsulation
You don't... really want Jolt types flying around your project.
